requirements.txt
flask
pillow
pytesseract
pdf2image
sentence-transformers
transformers
peft
redis
werkzeug
google-generativeai
numpy
torch
bitsandbytes
pdfplumber
pypdfium2
pdfminer.six
python-docx
datasets
sentencepiece
huggingface_hub
python-dotenv
flask_sqlalchemy
flask_login
flask_mail
flask_dance
google-auth
google-api-python-client